+++ CHANGELOG.md
@@ -18,4 +18,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 [0.1.0]: https://github.com/magnetcoop/hydrogen.module.session.keycloak/releases/tag/v0.1.0
 [0.1.1]: https://github.com/magnetcoop/hydrogen.module.session.keycloak/compare/v0.1.0...v0.1.1/
-[UNRELEASED]: https://github.com/magnetcoop/hydrogen.module.session.keycloak/compare/v0.1.0...HEAD/
+[UNRELEASED]: https://github.com/magnetcoop/hydrogen.module.session.keycloak/compare/v0.1.1...HEAD/