+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+
+##[0.1.2]
 ### Added
 - tests
 - Now you can configure this module to specify where does the necessary data come from.
@@ -24,4 +26,5 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 [0.1.0]: https://github.com/magnetcoop/hydrogen.module.session.keycloak/releases/tag/v0.1.0
 [0.1.1]: https://github.com/magnetcoop/hydrogen.module.session.keycloak/compare/v0.1.0...v0.1.1/
-[UNRELEASED]: https://github.com/magnetcoop/hydrogen.module.session.keycloak/compare/v0.1.1...HEAD/
+[0.1.2]: https://github.com/magnetcoop/hydrogen.module.session.keycloak/compare/v0.1.1...v0.1.2/
+[UNRELEASED]: https://github.com/magnetcoop/hydrogen.module.session.keycloak/compare/v0.1.2...HEAD/